+++ src/Sulu/Bundle/AutomationBundle/Resources/public/js/main.js
@@ -1,0 +1,27 @@
+/*
+ * This file is part of the Sulu CMS.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+require.config({
+    paths: {
+        suluautomation: '../../suluautomation/js'
+    }
+});
+
+define(function() {
+    return {
+
+        name: "SuluAutomationBundle",
+
+        initialize: function(app) {
+
+            'use strict';
+
+        }
+    };
+});