+++ src/Sulu/Bundle/AutomationBundle/SuluAutomationBundle.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Bundle\AutomationBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * Entry point for this bundle.
+ */
+class SuluAutomationBundle extends Bundle
+{
+}