+++ src/Sulu/Bundle/AutomationBundle/Tests/app/AppKernel.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+use Sulu\Bundle\TestBundle\Kernel\SuluTestKernel;
+
+/**
+ * Test kernel.
+ */
+class AppKernel extends SuluTestKernel
+{
+}